build/kernel: 11.2 does not write empty debug set

+++ build/kernel.sh
@@ -70,7 +70,7 @@ ${ENV_FILTER} make -s -C${SRCDIR}/release kernel.txz ${MAKE_ARGS}
 
 sh ./clean.sh ${SELF}
 
-if [ -z "$(tar -tf ${DEBUG_OBJ})" ]; then
+if [ -z "$(test -f ${DEBUG_OBJ} && tar -tf ${DEBUG_OBJ})" ]; then
 	echo -n ">>> Copying release kernel set... "
 	mv ${KERNEL_OBJ} ${KERNEL_RELEASE_SET}
 	echo "done"
